Put all creating operations in a seperate file
